[added] creature static spawn mechanics to lua
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk@3234 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ MMOCoreORB/src/server/zone/objects/creature/commands/DeclareresidenceCommand.h
@@ -88,7 +88,7 @@ public:
 			Time* cooldownTime = creature->getCooldownTime("declareresidence");
 
 			if (cooldownTime->isFuture()) {
-				int hours = (int) round((float) abs(cooldownTime->miliDifference()) / 3600000.f); //Divided by 1 hour.
+				int hours = (int) round((float) abs((float) cooldownTime->miliDifference()) / 3600000.f); //Divided by 1 hour.
 
 				StringIdChatParameter param("@player_structure:change_residence_time"); //You cannot change residence for %NO hours)
 				param.setTO(String::valueOf(hours));
